AcceptOrDeny now can take 3 parameters to update the pending_Bet

+++ Controllers/BetsController.cs
@@ -148,7 +148,7 @@ namespace SourDuckWannaBet.Controllers
             }
         }
 
-        public async Task<bool> AcceptOrDenyBetAsync(long betId, string newStatus, long pendingBet)
+        public async Task<bool> AcceptOrDenyBetAsync(long betId, string newStatus, double pendingBet)
         {
             try
             {
